# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ModelHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List all configured providers
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createProvider
      tags: [providers]
      summary: Create a new provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderInput"
      responses:
        "201":
          description: Created provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /providers/{id}:
    get:
      operationId: getProvider
      tags: [providers]
      summary: Get a provider by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProvider
      tags: [providers]
      summary: Update a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderPatch"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProvider
      tags: [providers]
      summary: Delete a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /providers/{id}/models:
    get:
      operationId: getProviderModels
      tags: [providers]
      summary: Fetch available models for a provider (from its API)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Models list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoveredModel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /providers/{id}/test:
    post:
      operationId: testProvider
      tags: [providers]
      summary: Test provider connectivity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- Detection ---
  /detect:
    post:
      operationId: detectProvider
      tags: [detection]
      summary: Auto-detect provider from API key, base URL, or endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectionInput"
      responses:
        "200":
          description: Detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectionResult"

  # --- Models ---
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List all saved models
      parameters:
        - name: providerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
    post:
      operationId: createModel
      tags: [models]
      summary: Save a model configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelInput"
      responses:
        "201":
          description: Created model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # --- Provider ---
    Provider:
      type: object
      required: [id, name, providerType, baseUrl, enabled, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        providerType:
          type: string
        apiKey:
          type: ["string", "null"]
        baseUrl:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        customHeaders:
          type: ["string", "null"]
        organizationId:
          type: ["string", "null"]
        endpoint:
          type: ["string", "null"]
        enabled:
          type: boolean
        detectedType:
          type: ["string", "null"]
        confidence:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProviderInput:
      type: object
      required: [name, providerType]
      properties:
        name:
          type: string
          minLength: 1
        providerType:
          type: string
        apiKey:
          type: string
        baseUrl:
          type: string
        model:
          type: string
        customHeaders:
          type: string
        organizationId:
          type: string
        endpoint:
          type: string
        enabled:
          type: boolean
    ProviderPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        providerType:
          type: string
        apiKey:
          type: string
        baseUrl:
          type: string
        model:
          type: string
        customHeaders:
          type: string
        organizationId:
          type: string
        endpoint:
          type: string
        enabled:
          type: boolean

    # --- Detection ---
    DetectionInput:
      type: object
      properties:
        apiKey:
          type: string
        baseUrl:
          type: string
        endpoint:
          type: string
    DetectionResult:
      type: object
      required: [detected, confidence]
      properties:
        detected:
          type: boolean
        providerType:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        confidence:
          type: integer
        suggestedBaseUrl:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
    TestResult:
      type: object
      required: [success, latencyMs]
      properties:
        success:
          type: boolean
        latencyMs:
          type: integer
        message:
          type: ["string", "null"]
        modelsEndpointWorking:
          type: boolean

    # --- Discovered Model (from provider API) ---
    DiscoveredModel:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        contextLength:
          type: ["integer", "null"]
        supportsVision:
          type: boolean
        supportsReasoning:
          type: boolean
        supportsFunctionCalling:
          type: boolean
        supportsImageGeneration:
          type: boolean
        supportsAudio:
          type: boolean
        supportsEmbeddings:
          type: boolean
        supportsStreaming:
          type: boolean
        supportsJsonMode:
          type: boolean

    # --- Saved Model ---
    Model:
      type: object
      required: [id, providerId, modelId, name, createdAt]
      properties:
        id:
          type: integer
        providerId:
          type: integer
        modelId:
          type: string
        name:
          type: string
        contextLength:
          type: ["integer", "null"]
        supportsVision:
          type: boolean
        supportsReasoning:
          type: boolean
        supportsFunctionCalling:
          type: boolean
        supportsImageGeneration:
          type: boolean
        supportsAudio:
          type: boolean
        supportsEmbeddings:
          type: boolean
        supportsStreaming:
          type: boolean
        supportsJsonMode:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ModelInput:
      type: object
      required: [providerId, modelId, name]
      properties:
        providerId:
          type: integer
        modelId:
          type: string
        name:
          type: string
        contextLength:
          type: integer
        supportsVision:
          type: boolean
        supportsReasoning:
          type: boolean
        supportsFunctionCalling:
          type: boolean
        supportsImageGeneration:
          type: boolean
        supportsAudio:
          type: boolean
        supportsEmbeddings:
          type: boolean
        supportsStreaming:
          type: boolean
        supportsJsonMode:
          type: boolean
...