# lib/api-zod/src/generated/types/detectionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ModelHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DetectionInput {
  apiKey?: string;
  baseUrl?: string;
  endpoint?: string;
}