# lib/api-zod/src/generated/types/detectionResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ModelHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DetectionResult {
  detected: boolean;
  /** @nullable */
  providerType?: string | null;
  /** @nullable */
  name?: string | null;
  confidence: number;
  /** @nullable */
  suggestedBaseUrl?: string | null;
  /** @nullable */
  reason?: string | null;
}