# lib/api-zod/src/generated/types/discoveredModel.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ModelHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiscoveredModel {
  id: string;
  name: string;
  /** @nullable */
  contextLength?: number | null;
  supportsVision?: boolean;
  supportsReasoning?: boolean;
  supportsFunctionCalling?: boolean;
  supportsImageGeneration?: boolean;
  supportsAudio?: boolean;
  supportsEmbeddings?: boolean;
  supportsStreaming?: boolean;
  supportsJsonMode?: boolean;
}