# lib/api-zod/src/generated/types/providerPatch.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ModelHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderPatch {
  /** @minLength 1 */
  name?: string;
  providerType?: string;
  apiKey?: string;
  baseUrl?: string;
  model?: string;
  customHeaders?: string;
  organizationId?: string;
  endpoint?: string;
  enabled?: boolean;
}